Update cmdlet names.
