+++ src/Commands/Site/NewSiteFileVersionBatchDeleteJob.cs
@@ -7,15 +7,11 @@ using System.Management.Automation;
 
 using Resources = PnP.PowerShell.Commands.Properties.Resources;
 
-namespace PnP.PowerShell.Commands.Lists
+namespace PnP.PowerShell.Commands.Sites
 {
-    [Cmdlet(VerbsCommon.New, "LibraryLevelFileVersionBatchDeleteJob")]
-    public class NewLibraryLevelFileVersionBatchDeleteJob : PnPWebCmdlet
+    [Cmdlet(VerbsCommon.New, "PnPSiteFileVersionBatchDeleteJob")]
+    public class NewSiteFileVersionBatchDeleteJob : PnPSharePointCmdlet
     {
-        [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0)]        
-        [ValidateNotNull]
-        public ListPipeBind Identity;
-
         [Parameter(Mandatory = true)]
         public int DeleteBeforeDays;
 
@@ -26,14 +22,11 @@ namespace PnP.PowerShell.Commands.Lists
         {
             if (Force || ShouldContinue("By executing this command, versions specified will be permanently deleted. These versions cannot be restored from the recycle bin. Are you sure you want to continue?", Resources.Confirm))
             {
-                var list = Identity.GetList(CurrentWeb);
-                if (list != null)
-                {
-                    list.StartDeleteFileVersions(DeleteBeforeDays);
-                    ClientContext.ExecuteQueryRetry();
-                    
-                    WriteVerbose("Success. Versions specified will be permanently deleted in the upcoming days.");
-                }
+                var site = ClientContext.Site;
+                site.StartDeleteFileVersions(DeleteBeforeDays);
+                ClientContext.ExecuteQueryRetry();
+
+                WriteVerbose("Success. Versions specified will be permanently deleted in the upcoming days.");
             }
             else
             {
